refactor(broker): narrow public crate API
