+++ .trajectories/completed/2026-05/traj_zqwco4gl76g3.json
@@ -1,0 +1,65 @@
+{
+  "id": "traj_zqwco4gl76g3",
+  "version": 1,
+  "task": {
+    "title": "Fix issue 878"
+  },
+  "status": "completed",
+  "startedAt": "2026-05-19T04:18:25.024Z",
+  "completedAt": "2026-05-19T04:27:18.903Z",
+  "agents": [
+    {
+      "name": "default",
+      "role": "lead",
+      "joinedAt": "2026-05-19T04:22:25.398Z"
+    }
+  ],
+  "chapters": [
+    {
+      "id": "chap_5yldelefew3p",
+      "title": "Work",
+      "agentName": "default",
+      "startedAt": "2026-05-19T04:22:25.398Z",
+      "endedAt": "2026-05-19T04:27:18.903Z",
+      "events": [
+        {
+          "ts": 1779164545398,
+          "type": "decision",
+          "content": "Route broker binary through library entry point: Route broker binary through library entry point",
+          "raw": {
+            "question": "Route broker binary through library entry point",
+            "chosen": "Route broker binary through library entry point",
+            "alternatives": [],
+            "reasoning": "Moving main.rs to call relay_broker::run_cli lets implementation modules live in the library crate and become pub(crate) instead of public Rust API while preserving the binary behavior."
+          },
+          "significance": "high"
+        },
+        {
+          "ts": 1779164631929,
+          "type": "decision",
+          "content": "Keep protocol and snippets as public Rust modules: Keep protocol and snippets as public Rust modules",
+          "raw": {
+            "question": "Keep protocol and snippets as public Rust modules",
+            "chosen": "Keep protocol and snippets as public Rust modules",
+            "alternatives": [],
+            "reasoning": "Current external Rust references only need snippets and the wire protocol remains the crate's intentional stable surface; broker runtime, relaycast plumbing, PTY, scheduling, metrics, and worker internals can be crate-private."
+          },
+          "significance": "high"
+        }
+      ]
+    }
+  ],
+  "retrospective": {
+    "summary": "Narrowed the broker Rust crate public API to protocol/snippets/run_cli, moved the binary entry through the library, colocated broker and worker tests, documented the Rust API break, and verified cargo fmt, cargo clippy -D warnings, and cargo test --release.",
+    "approach": "Standard approach",
+    "confidence": 0.9
+  },
+  "commits": [],
+  "filesChanged": [],
+  "projectId": "/Users/will/Projects/AgentWorkforce/relay",
+  "tags": [],
+  "_trace": {
+    "startRef": "c54e118806e98b9defe0e0b5022c35be8e64a52f",
+    "endRef": "c54e118806e98b9defe0e0b5022c35be8e64a52f"
+  }
+}
