+++ .trajectories/index.json
@@ -1,6 +1,6 @@
 {
   "version": 1,
-  "lastUpdated": "2026-05-19T03:54:07.051Z",
+  "lastUpdated": "2026-05-19T04:27:19.071Z",
   "trajectories": {
     "traj_05xg7j388bc4": {
       "title": "Add browser workflow step integration",
@@ -995,6 +995,13 @@
       "startedAt": "2026-05-19T03:40:40.798Z",
       "completedAt": "2026-05-19T03:54:06.889Z",
       "path": "/Users/will/Projects/AgentWorkforce/relay/.trajectories/completed/2026-05/traj_5qbla7w4kzoi.json"
+    },
+    "traj_zqwco4gl76g3": {
+      "title": "Fix issue 878",
+      "status": "completed",
+      "startedAt": "2026-05-19T04:18:25.024Z",
+      "completedAt": "2026-05-19T04:27:18.903Z",
+      "path": "/Users/will/Projects/AgentWorkforce/relay/.trajectories/completed/2026-05/traj_zqwco4gl76g3.json"
     }
   }
 }
