+++ crates/broker/src/lib.rs
@@ -1,25 +1,54 @@
-pub mod auth;
-pub mod config;
-pub mod control;
-pub mod conversation_log;
-pub mod crash_insights;
-pub mod dedup;
-pub mod events;
-pub mod inject;
-pub mod message_bridge;
-pub mod metrics;
-pub mod multi_workspace;
-pub mod priorities;
 pub mod protocol;
-pub mod pty;
-pub mod queue;
-pub mod redact;
-pub mod relaycast;
-pub mod relaycast_ws;
-pub mod replay_buffer;
-pub mod scheduler;
-pub mod snapshot;
 pub mod snippets;
-pub mod supervisor;
-pub mod telemetry;
-pub mod types;
+
+pub(crate) mod broker;
+pub(crate) mod cli;
+pub(crate) mod cli_mcp_args;
+#[allow(dead_code)]
+pub(crate) mod config;
+pub(crate) mod control;
+#[allow(dead_code)]
+pub(crate) mod conversation_log;
+pub(crate) mod crash_insights;
+#[allow(dead_code)]
+pub(crate) mod dedup;
+#[allow(dead_code)]
+pub(crate) mod events;
+pub(crate) mod listen_api;
+#[allow(dead_code)]
+pub(crate) mod metrics;
+pub(crate) mod priorities;
+#[allow(dead_code)]
+pub(crate) mod pty;
+pub(crate) mod pty_worker;
+#[allow(dead_code)]
+pub(crate) mod queue;
+pub(crate) mod readiness;
+#[allow(dead_code)]
+pub(crate) mod redact;
+#[allow(dead_code)]
+pub(crate) mod relaycast;
+pub(crate) mod replay_buffer;
+pub(crate) mod routing;
+pub(crate) mod runtime;
+#[allow(dead_code)]
+pub(crate) mod scheduler;
+pub(crate) mod snapshot;
+pub(crate) mod spawner;
+#[allow(dead_code)]
+pub(crate) mod supervisor;
+pub(crate) mod swarm;
+pub(crate) mod swarm_tui;
+#[allow(dead_code)]
+pub(crate) mod telemetry;
+#[allow(dead_code)]
+pub(crate) mod types;
+pub(crate) mod util;
+pub(crate) mod wait;
+pub(crate) mod worker;
+pub(crate) mod worker_request;
+pub(crate) mod wrap;
+
+pub async fn run_cli() -> anyhow::Result<()> {
+    cli::run().await
+}
