+++ crates/broker/src/wait.rs
@@ -343,7 +343,7 @@ fn preferred_trigger(conds: &[WaitCondition]) -> Trigger {
 /// Pre-built text-only `WaitSet`s expressing each CLI's "screen looks
 /// ready" rule.
 #[allow(dead_code)]
-pub mod for_cli {
+pub(crate) mod for_cli {
     use std::time::Duration;
 
     use super::{WaitSet, IDLE_SETTLE};
