+++ crates/broker/src/worker.rs
@@ -5,13 +5,13 @@ use std::{
     time::{Duration, Instant},
 };
 
-use anyhow::{Context, Result};
-use relay_broker::{
+use crate::{
     metrics::MetricsCollector,
     protocol::{AgentRuntime, AgentSpec, ProtocolEnvelope, RelayDelivery, PROTOCOL_VERSION},
     relaycast::configure_relaycast_mcp_with_token,
     supervisor::Supervisor,
 };
+use anyhow::{Context, Result};
 use serde_json::{json, Value};
 use tokio::{
     io::{AsyncBufReadExt, AsyncWriteExt, BufReader},
@@ -945,6 +945,50 @@ fn spawn_worker_reader<R>(
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    fn make_registry(env: Vec<(String, String)>) -> WorkerRegistry {
+        let (tx, _rx) = mpsc::channel::<WorkerEvent>(16);
+        WorkerRegistry::new(tx, env, PathBuf::from("/tmp/worker-tests"), Instant::now())
+    }
+
+    #[test]
+    fn worker_registry_starts_empty() {
+        let reg = make_registry(vec![]);
+        assert!(!reg.has_any_worker());
+        assert!(reg.list().is_empty());
+    }
+
+    #[test]
+    fn has_worker_returns_false_for_unknown() {
+        let reg = make_registry(vec![]);
+        assert!(!reg.has_worker("nonexistent"));
+    }
+
+    #[test]
+    fn worker_log_path_rejects_path_traversal() {
+        let reg = make_registry(vec![]);
+        assert!(reg.worker_log_path("..").is_none());
+        assert!(reg.worker_log_path("../etc/passwd").is_none());
+        assert!(reg.worker_log_path("foo/../bar").is_none());
+        assert!(reg.worker_log_path("foo/bar").is_none());
+        assert!(reg.worker_log_path("foo\\bar").is_none());
+        assert!(reg.worker_log_path("valid-name").is_some());
+        assert!(reg.worker_log_path("worker.1").is_some());
+    }
+
+    #[test]
+    fn env_value_lookup() {
+        let env = vec![("KEY".into(), "val".into())];
+        let reg = make_registry(env);
+        assert_eq!(reg.env_value("KEY"), Some("val"));
+        assert_eq!(reg.env_value("MISSING"), None);
+    }
+
+    #[test]
+    fn routing_workers_empty_when_no_workers() {
+        let reg = make_registry(vec![]);
+        assert!(reg.routing_workers().is_empty());
+    }
 
     #[test]
     fn args_include_model_override_detects_supported_forms() {
